reduced n_estimators to 2
