+++ dashboard/app.py
@@ -10,10 +10,8 @@ import plotly.graph_objs as go
 import pandas_datareader.data as web
 from sklearn.metrics import mean_squared_error
 import xgboost as xgb
-from models import xgb_price_predictor
-
-# CONSTANTS
-WINDOW_SIZE = 2
+# from .models import xgbr_predict
+from xgb_model import xgb_price_predictor
 
 # Get data
 start = datetime.datetime(2014,9,15)
@@ -45,7 +43,7 @@ train_starts = {
 app = dash.Dash(__name__, 
     external_stylesheets=[dbc.themes.BOOTSTRAP], 
     meta_tags=[{'name':'viewport', 'content': 'width=device-width, initial-scale=1.0'}],
-    title="crypto-price-prediction-ml")
+    title="dash-app-layout")
 server = app.server
 
 # components
@@ -58,7 +56,7 @@ header = html.Div(
     ], className="main-header mb-4 py-3 text-center"
 )
 
-# historical and prediction
+# first row
 div_0 = html.Div([
     dbc.Row([
         dbc.Col(html.Div([
@@ -67,9 +65,15 @@ div_0 = html.Div([
         dbc.Col(html.Div([
             html.H5("83.2%"),
             html.P("Model Test Accuracy")],className="model-info-item py-2")),
+        # dbc.Col(html.Div([
+        #     html.H3("70/15/15"),
+        #     html.P("Training/Test/Validation Split")], className="model-info-item py-2")),
     ], className="model-info my-3 py-2 text-center")
     ])
 
+
+
+# second row
 div_1 = html.Div([
     dbc.Row([
         dbc.Col(html.Div([
@@ -85,6 +89,11 @@ div_1 = html.Div([
                     html.P(id="nextday"),
                     html.Div(id="pred-price", className="nextday-price text-center my-1 px-3 py-1")
                     ], className="nextday-col text-center"),
+                # dbc.Col([
+                #     html.H5("Next Week Avg.", className="pred-price"),
+                #     html.Div(id="nextweek-avg", className="nextday-price text-center my-1 px-3 py-1")
+                #     ], className="nextday-col text-center"),
+                # dbc.Col(html.H4("Predicted Price", className="pred-price")),
                 ]),
             dbc.Row([
                 dbc.Col(dcc.Graph(id="pred-plot", figure={}))
@@ -98,6 +107,11 @@ div_2 = html.Div([
     dbc.Row([
         dbc.Col([
             html.H3("Weekly Volatility"),
+            # dcc.Checklist(
+            #     id="btc-vol",
+            #     options=["BTC Volatility"],
+            #     value=["btc-voltaility"]
+            #     ),
             dcc.Graph(id="volatility", figure={})
             ])      
         ], className="volt-div mb-4")
@@ -108,6 +122,7 @@ app.layout = dbc.Container([
     header,
     div_0,
     div_1,
+    # html.Br(),
     div_2,
     ],className="page-container", fluid=True)
 
@@ -115,7 +130,9 @@ app.layout = dbc.Container([
 # =========
 # callbacks
 # =========
-
+# 
+# 
+# 
 # time series plot
 @app.callback(
     Output("tseries-fig", "figure"),
@@ -125,9 +142,9 @@ app.layout = dbc.Container([
 def update_graph(coin_selected):
     dff = df[df["Symbols"] == coin_selected]
     # simple moving average using pandas rolling window
-    dff["SMA_6Month"] = dff["Close"].copy().rolling(window=180).mean()
+    dff["SMA_6Month"] = dff["Close"].rolling(window=180).mean()
     # cumulative moving average
-    dff["CMA"] = dff["Close"].copy().expanding(min_periods=3).mean()
+    dff["CMA"] = dff["Close"].expanding(min_periods=3).mean()
     fig_1 = px.line(dff, x="Date", y=["Close", "SMA_6Month", "CMA"], title="Historical Prices",
          width=500, height=400,
         labels={"Date": "", "Value": "Coin Value (USD)"}
@@ -154,6 +171,7 @@ def update_graph(coin_selected):
     Output("pred-plot", "figure"),
     Output("pred-price", "children"),
     Output("nextday", "children"),
+    # Output("nextweek-avg", "children"),
     Input("select-coin", "value")
     )
 def update_nextday_price(coin_selected):
@@ -166,9 +184,9 @@ def update_nextday_price(coin_selected):
     df1.set_index("Date", inplace=True)
     df1 = df1[["Close"]].copy()
     train_date = train_starts[coin_selected]
-
-    plot_df, pred_price = xgb_price_predictor(df1, train_date, WINDOW_SIZE)
-    fig_pred = px.line(plot_df, y=["Close", "Prediction"], title="Actual vs Predicted Close Prices",
+    # plot_df, pred_price = xgbr_predict(df1, train_date, 6)
+    plot_df, pred_price = xgb_price_predictor(df1, train_date, 6)
+    fig_pred = px.line(plot_df, y=["Close", "Pred"], title="Actual vs Predicted",
         width=600, height=400,
         labels={"Date": "", "value": "Coin Value (USD)"}
         )
@@ -180,6 +198,7 @@ def update_nextday_price(coin_selected):
         color="RebeccaPurple"
     ),
     title=dict(
+        # text="Time Series",
         x=0.5,
         y=0.9,
         xanchor='center',
@@ -195,20 +214,24 @@ def update_nextday_price(coin_selected):
     )
     fig_pred.update_layout(
         margin=dict(l=20, r=20, t=20, b=20),
+        # paper_bgcolor="LightSteelBlue",
         )
     fig_pred.update_xaxes(tickangle=-45)
 
-    return fig_pred, f"${pred_price:.2f}", f"({tomorrow})"
+    return fig_pred, f"${pred_price:.0f}", f"({tomorrow})"
 
 # weekly volatility
 @app.callback(
     Output("volatility", "figure"),
+    # Output("btc-vol", "figure"),
     Input("select-coin", "value")
     )
 def update_volatility_graph(coin_selected):
+    btc_df = df[df["Symbols"] == "BTC-USD"]
     dff = df[df["Symbols"] == coin_selected]
     dff["weekly_volt"] = dff["Close"].pct_change().rolling(7).std()
-    fig_3 = px.line(dff, x="Date", y="weekly_volt", title="Weekly Volatility",
+    btc_df["weekly_volt"] = btc_df["Close"].pct_change().rolling(7).std()
+    fig_3 = px.line(dff, x="Date", y="weekly_volt", title="Weekly Close Price Volatility",
         labels={"Date": "", "weekly_volt": "Weekly Volatility"}
         )
     # update
@@ -218,10 +241,11 @@ def update_volatility_graph(coin_selected):
         size=13,
     ),
     title=dict(
-        x=0.5,
-        y=0.9,
-        xanchor='center',
-        yanchor= 'top',
+        # # text="Time Series",
+        # x=0.3,
+        # y=0.9,
+        # xanchor='center',
+        # yanchor= 'top',
         font=dict(
             family="Open Sans",
             color="orange",
@@ -232,10 +256,6 @@ def update_volatility_graph(coin_selected):
     fig_3.update_xaxes(tickangle=-45)
 
     return fig_3
-
-
-
 # main
-# ====
 if __name__ == '__main__':
     app.run_server(debug=True, port=3342)
