+++ dashboard/models.py
@@ -1,80 +1,62 @@
+import datetime
 import pandas as pd
-import xgboost as xgb
+from xgboost import XGBRegressor
 
-XGBR_N_ESTIMATORS = 10
+XGBR_N_ESTIMATORS = 2
 LEARNING_RATE = 0.1
-# nextday price prediction function
-# =================================
-def xgb_price_predictor(data, training_end_date, window_size):
-    # create features
-    df = create_features(data, window_size)
+#
+today = datetime.date.today()
+tomorrow = today + datetime.timedelta(days=1)
+
+
+# create function to use XGBRegressor for next day prediction
+def xgbr_predict(data_df, train_start_date, window_size=6):
+    """This module will use the Closing priice of a cryptocurrency 
+    to predict the next day price using XGBRegressor.
+
+    INPUTS: 
+    	- data_df: Historical data from yfinance
+    	- train_strat_dat: date used to split data into training and testing sets
+    	- window_size: how many past days to consider for a one-step prediction
+
+    OUTPUT: 
+    	- plot_df: a data frame with the actual closing prices used in the testing 
+    	and predicted closing prices
+    	- pred[-1] the predicted value for next day
+    """
+    df = data_df[["Close"]].copy()
+    n = train_start_date
+
+    for i in range(1, window_size+1):
+        df[f'Close-{i}'] = df['Close'].shift(i)
+
     
-    # Training / Test split
-    X_train, y_train, X_test, y_test, test_date_index = tss(df, training_end_date)
-    
-    
-    # train model
-    model = xgb.XGBRegressor(objective="reg:squarederror",
-                         n_estimators=XGBR_N_ESTIMATORS,
-                         learning_rate=LEARNING_RATE,
-                         random_state=1,
-                         # early_stoping_rounds=50
-                        )
+    df = df.dropna()
+    # features and target 
+    y = df[["Close"]]
+    X = df.drop(columns=["Close"])
+    # train test split
+    X_train = X.loc[:n]
+    y_train = y.loc[:n]
+    X_test = X.loc[n:]
+    y_test = y.loc[n:]
+
+    # train and fit model
+    model = XGBRegressor(
+        objective="reg:squarederror", 
+        n_estimators=XGBR_N_ESTIMATORS, 
+        learning_rate=LEARNING_RATE)
+
     model.fit(X_train, y_train)
-    
-    # make prediction on the Test set
-    y_pred = model.predict(X_test)
-    
-    # create plot_df 
-    plot_df = pd.DataFrame(y_test, columns=["Close"])
-    plot_df["Prediction"] = y_pred
-    plot_df.index = test_date_index
-    
-    # make nextday prediction
-    X_today = y_test[-window_size::][::-1].reshape(1, window_size)    
-    y_nextday = model.predict(X_today)
-    
-    # return test/prediction plot and nextday prediction
-    return plot_df, y_nextday[0]
+    # predict
+    pred = model.predict(X_test)
+    # make plot df
+    plot_df = pd.DataFrame(y_test)
+    plot_df["Pred"] = pred
+    plot_df.index = y_test.index
 
+    # test0 = test.loc["2022-08-08"]
 
-# create features
-# ===============
-def create_features(data, window_size):
-    '''
-    This function converts the "Close" price column into a features table
-    using a given window size
-    
-    INPUTS:
-    ------
-    data: the row data from Yahoo Finance
-    window_size:
-    
-    OUTPUT:
-    ------
-    df: a data frame with the "Close"
-    '''
-    df = data[["Close"]].copy()
-    
-    for i in range(1, window_size + 1):
-        df[f"Close-{i}"] = df["Close"].shift(i)
-        
-    df.dropna(inplace=True)
-    
-    return df
+    # return plot_df and the last predicted value, pred[-1]
+    return plot_df, pred[-1]
 
-# training/test split
-# ===================
-def tss(features_df, training_end_date):
-    df = features_df.copy()
-    n = training_end_date
-    train = df.loc[:n]
-    test = df.loc[n:]
-    X_train = train.drop(columns="Close").values
-    y_train = train[["Close"]].values
-    X_test = test.drop(columns="Close").values
-    y_test = test[["Close"]].values
-    test_date_index = test.index
-    
-    return X_train, y_train, X_test, y_test, test_date_index
-#==================== END ==================================
